+++ src/modules/postcomments.js
@@ -1,3 +1,4 @@
+import createcomment from "./createComment.js";
 
 export default async function postcoment(e){
         const username = document.getElementById('username').value;
@@ -19,24 +20,10 @@ export default async function postcoment(e){
                 body: JSON.stringify(data),
               },
             );
-            const result = await responst.text()
-            console.log(result);
+            createcomment(item);
+            const result = await responst.text();
+             return result;
         }
-        const urlgetcomment=`https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/OfV1qInwl9XDUHwiPKmO/comments?item_id=${item}`
-          const infocoment = await fetch(urlgetcomment,{
-              method: 'GET',
-              headers: {
-                'Content-type': 'application/json; charset=UTF-8',
-              },
-            },
-          );
-          const infocmnt = await infocoment.json();
-          console.log(infocmnt)
-          const displaycoments= document.getElementById('displaycomments');
-          displaycoments.innerHTML=''; 
-          infocmnt.forEach((item) => {
-            const showcomments =`<li>${item.creation_date} : ${item.username} :${item.comment} </li>`;
-            displaycoments.insertAdjacentHTML('beforeend', showcomments);
-            });
-
+    
 }
+
